Add collage generation based on groq + google vertex api sentence embeddings and refactor python service greatly for file separation
